3rd, 4th, 5th and 6th days solutions.

+++ day_template.rs
@@ -1,8 +1,8 @@
-fn solve_part_1(input: &str) -> i32 {
+fn solve_part_1(input: &str) -> u64 {
     todo!();
 }
 
-fn solve_part_2(input: &str) -> i32 {
+fn solve_part_2(input: &str) -> u64 {
     todo!();
 }
